+++ knockback.js
@@ -1015,7 +1015,7 @@ return /******/ (function(modules) { // webpackBootstrap
 /* 5 */
 /***/ function(module, exports, __webpack_require__) {
 
-	
+	/* WEBPACK VAR INJECTION */(function(global) {
 	/*
 	  knockback.js 0.18.6
 	  Copyright (c)  2011-2014 Kevin Malakoff.
@@ -1024,7 +1024,9 @@ return /******/ (function(modules) { // webpackBootstrap
 	  Dependencies: Knockout.js, Backbone.js, and Underscore.js (or LoDash.js).
 	  Optional dependencies: Backbone.ModelRef.js and BackboneORM.
 	 */
-	var Backbone, copyProps, kb, ko, _;
+	var Backbone, copyProps, kb, ko, root, _;
+
+	root = typeof window !== "undefined" && window !== null ? window : global;
 
 	ko = __webpack_require__(18);
 
@@ -1290,9 +1292,9 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	})();
 
-	if (typeof window !== "undefined" && window !== null ? window.Parse : void 0) {
-	  Backbone = kb.Parse = window.Parse;
-	  _ = kb._ = window.Parse._;
+	if (root.Parse) {
+	  Backbone = kb.Parse = root.Parse;
+	  _ = kb._ = root.Parse._;
 	} else {
 	  Backbone = kb.Backbone = __webpack_require__(19);
 	  _ = kb._ = __webpack_require__(20);
@@ -1306,8 +1308,9 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	kb.Events = Backbone.Events;
 
-	kb.$ = (typeof window !== "undefined" && window !== null ? window.jQuery : void 0) || (typeof window !== "undefined" && window !== null ? window.$ : void 0);
-
+	kb.$ = root.jQuery || root.$;
+	
+	/* WEBPACK VAR INJECTION */}.call(exports, (function() { return this; }())))
 
 /***/ },
 /* 6 */
@@ -1565,7 +1568,7 @@ return /******/ (function(modules) { // webpackBootstrap
 /* 7 */
 /***/ function(module, exports, __webpack_require__) {
 
-	
+	/* WEBPACK VAR INJECTION */(function(global) {
 	/*
 	  knockback.js 0.18.6
 	  Copyright (c)  2011-2014 Kevin Malakoff.
@@ -1574,7 +1577,9 @@ return /******/ (function(modules) { // webpackBootstrap
 	  Dependencies: Knockout.js, Backbone.js, and Underscore.js (or LoDash.js).
 	  Optional dependencies: Backbone.ModelRef.js and BackboneORM.
 	 */
-	var ORM, ORMAdapter_BackboneAssociations, ORMAdapter_BackboneRelational, ORMAdapter_Supermodel, kb, _;
+	var AssociatedModel, ORM, ORMAdapter_BackboneAssociations, ORMAdapter_BackboneRelational, ORMAdapter_Supermodel, RelationalModel, Supermodel, kb, root, _;
+
+	root = typeof window !== "undefined" && window !== null ? window : global;
 
 	kb = __webpack_require__(5);
 
@@ -1632,17 +1637,19 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	kb.orm = new ORM();
 
+	RelationalModel = null;
+
 	ORMAdapter_BackboneRelational = (function() {
 	  function ORMAdapter_BackboneRelational() {}
 
 	  ORMAdapter_BackboneRelational.prototype.isAvailable = function() {
 	    var _ref;
-	    return !!((_ref = kb.Backbone) != null ? _ref.RelationalModel : void 0);
+	    return !!(RelationalModel = (_ref = kb.Backbone) != null ? _ref.RelationalModel : void 0);
 	  };
 
 	  ORMAdapter_BackboneRelational.prototype.relationType = function(model, key) {
 	    var relation;
-	    if (!(model instanceof kb.Backbone.RelationalModel)) {
+	    if (!(model instanceof RelationalModel)) {
 	      return null;
 	    }
 	    if (!(relation = _.find(model.getRelations(), function(test) {
@@ -1699,16 +1706,18 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	kb.orm.addAdapter(new ORMAdapter_BackboneRelational());
 
+	AssociatedModel = null;
+
 	ORMAdapter_BackboneAssociations = (function() {
 	  function ORMAdapter_BackboneAssociations() {}
 
 	  ORMAdapter_BackboneAssociations.prototype.isAvailable = function() {
 	    var _ref;
-	    return !!((_ref = kb.Backbone) != null ? _ref.AssociatedModel : void 0);
+	    return !!(AssociatedModel = (_ref = kb.Backbone) != null ? _ref.AssociatedModel : void 0);
 	  };
 
 	  ORMAdapter_BackboneAssociations.prototype.keys = function(model) {
-	    if (!(model instanceof kb.Backbone.AssociatedModel)) {
+	    if (!(model instanceof AssociatedModel)) {
 	      return null;
 	    }
 	    return _.map(model.relations, function(test) {
@@ -1718,7 +1727,7 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	  ORMAdapter_BackboneAssociations.prototype.relationType = function(model, key) {
 	    var relation;
-	    if (!(model instanceof kb.Backbone.AssociatedModel)) {
+	    if (!(model instanceof AssociatedModel)) {
 	      return null;
 	    }
 	    if (!(relation = _.find(model.relations, function(test) {
@@ -1739,11 +1748,13 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	kb.orm.addAdapter(new ORMAdapter_BackboneAssociations());
 
+	Supermodel = null;
+
 	ORMAdapter_Supermodel = (function() {
 	  function ORMAdapter_Supermodel() {}
 
 	  ORMAdapter_Supermodel.prototype.isAvailable = function() {
-	    return !!(typeof window !== "undefined" && window !== null ? window.Supermodel : void 0);
+	    return !!(Supermodel = root.Supermodel);
 	  };
 
 	  ORMAdapter_Supermodel.prototype.keys = function(model) {
@@ -1804,7 +1815,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	})();
 
 	kb.orm.addAdapter(new ORMAdapter_Supermodel());
-
+	
+	/* WEBPACK VAR INJECTION */}.call(exports, (function() { return this; }())))
 
 /***/ },
 /* 8 */
@@ -2635,7 +2647,7 @@ return /******/ (function(modules) { // webpackBootstrap
 /* 12 */
 /***/ function(module, exports, __webpack_require__) {
 
-	
+	/* WEBPACK VAR INJECTION */(function(global) {
 	/*
 	  knockback.js 0.18.6
 	  Copyright (c)  2011-2014 Kevin Malakoff.
@@ -2644,9 +2656,13 @@ return /******/ (function(modules) { // webpackBootstrap
 	  Dependencies: Knockout.js, Backbone.js, and Underscore.js (or LoDash.js).
 	  Optional dependencies: Backbone.ModelRef.js and BackboneORM.
 	 */
-	var kb;
+	var kb, root, _;
+
+	root = typeof window !== "undefined" && window !== null ? window : global;
 
 	module.exports = kb = __webpack_require__(5);
+
+	_ = kb._;
 
 	kb.modules = {
 	  underscore: kb._,
@@ -2654,17 +2670,9 @@ return /******/ (function(modules) { // webpackBootstrap
 	  knockout: kb.ko
 	};
 
-	if (typeof window !== "undefined" && window !== null) {
-	  window._ || (window._ = kb._);
-	  if (kb.Backbone) {
-	    window.Backbone || (window.Backbone = kb.Backbone);
-	  }
-	  if (kb.Parse) {
-	    window.Parse || (window.Parse = kb.Parse);
-	  }
-	  window.ko || (window.ko = kb.ko);
-	}
-
+	_.defaults(root, _.pick(kb, '_', 'Backbone', 'Parse', 'ko', '$'));
+	
+	/* WEBPACK VAR INJECTION */}.call(exports, (function() { return this; }())))
 
 /***/ },
 /* 13 */