+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Please refer to the following release notes when upgrading your version of Knock
 ## 0.19.0
 * Removed standalone plugins. Just choose knockback if you need one or more plugins.
 * Change library building to use webpack.
+   - Webpack doesn't support optional dependencies well so for you should manually load jquery or Supermodel onto globals in Node.js, for example: global.jQuery = require 'jquery'
 * Avoid creating dependencies in collectionObservable's destroy: https://github.com/kmalakoff/knockback/pull/120
 
 ## 0.18.6